[FLINK-15818] [kafka-io] Let ingress builder accept deserializer instance for runtime usage
This is a preliminary change to unify source creation logic between the
ProtobufKafkaSourceProvider and KafkaSourceProvider, so that YAML
configurations for a protobuf Kafka ingress can be parsed as a
KafkaIngressSpec.

We do not yet want to allow users to configure their Kafka ingress
deserializer as instances, so for now this is annotated with ForRuntime.

This commit also changes when we instantiate the deserializer.
Previously, this occurred within the KafkaSourceProvider. We now
instantiate the deserializer within the builder.