[FLINK-15839] [kafka-io] Use OptionalConfig in KafkaIngressBuilder to handle config resolution
With this commit, we now have the following common strategy for dealing
with optional configuration with or without defaults:

* All optional config fields should be an OptionalConfig
* For optional configs that should be resolved as a Kafka client
property, if the value is set by the user using a named config method,
then the set value always overwrites any value set in the explicitly
provided Properties.
* Otherwise, if a value was not set by the user but it has a default
value defined, we set the default value in the resolved Properties IFF
no value was set in the explicitly provided Properties.