[FLINK-16123] Move PROTOBUF_KAFKA_INGRESS_TYPE from SDK to statefun-flink-io
The PROTOBUF_KAFKA_INGRESS_TYPE shouldn't be something exposed through
the Java SDK, as it is only meant for usage by polyglot users who
defined their ingress using YAMLs.

This has been moved to statefun-flink-io, which is a common dependency
between statefun-flink-io-bundle and statefun-flink-core, serving as a
good basis for adding new Protobuf Kafka ingress types in the future.