[FLINK-16123] Add RoutableKafkaVerificationModule application for E2E testing
This is a a simple application used for testing the routable Kafka
ingress.

The application reads untagged messages from a Kafka ingress, binds
multiple functions which has the sole purpose of tagging of messages
with their own addresses (see FnSelfAddressTagger), and then sending
back the tagged messages back to a Kafka egress.