[FLINK-16123] [e2e] Add RoutableKafkaE2E
End-to-end test based on the {@link RoutableKafkaVerificationModule} application.

This test writes some records to Kafka, with target function id as key
(UTF8 String) and MessageWithAddress messages as value, without the
"from" field set. The routable Kafka ingress should automatically route
them to the correct function instances, which tag the input messages
with their own address, and then forwards it back to Kafka. The test
verifies that the tagged outputs written back to Kafka are correct.