[FLINK-16311] [sdk] Fix user contracts for PersistedAppendingBuffer
The previous contracts were not possible to be provided, since Flink
ListState never returns null when retrieving the list, regardless of if
the access was after a clear, or the list was actually just empty.

This commit changes the contract of the view() method to always return
non-null, and also disallow appending / update the buffer with null
elements or lists.