[FLINK-16838] Let Dockerfile in quickstart point to a specific image version tag
With this change, the quickstart archetype now always uses a specific
image version.

If the archetype is a released artifact, then the version tag used will
be x.y.z, e.g. 2.0.0. Images in this case should be available in the
Docker Hub registry.

Otherwise, if the archetype is a snapshot distribution built from
development branches, then the version tag used wil be snapshot
versions, e.g. 2.1-SNAPSHOT. In this case, you would have to locally
build the image from the corresponding source version.