[FLINK-16897][docs] Consolidate doc images in the "fig" directory.
Because binary image files cannot have a license header, we typically put them into a
folder with a "license" file that describes that the images in the folder are licensed
to ASF.

To follow that best practice, this commit moves the images in docs/images to docs/fig,
which contains such a file.

This closes #85.