[hotfix] [build] Fix path to spotbugs exclude filter file
This commit fixes the path so that it is absolute, instead of a relative
path to the root directory. This was causing issues, since developers
essentially could not build modules individually due to an error
complaining that the exclude file cannot be located (since the path was
relative to the root, and only works when building from root).