[FLINK-19001] Make HttpFunctionSpec java serializable
This commit makes the internal class HttpFunctionSpec java
serializable, so that it can be used from the statefun-datastream-api.
If at any point we will decide to add a non (trivally serializable) member
to this class, then we can undo this change and provide a serializable
"twin" at the datastream api.