[FLINK-19102] [core, sdk] Make StateBinder a per-FunctionType entity
The goal of this commit is to make StateBinders know the FunctionType,
with each state binder instance specifically binding state for a single
FunctionType.

The side effect of this is also moving FunctionType out of
PersistedStateRegistry, as that information is encapsulated to the
StateBinder it receives at runtime.

This closes #137.