[hotfix][dockerfile] Add newline to end of flink-conf
When building small applications it's often easy to simply append
flink configurations to the provided base image directly inside the
Dockerfile. echo does not prepend a newline so its convenient to leave
one at the end of the file. Otherwise, the first configuration is added
to the task manager memory amount which leads to a strange parse error on
deployment.