[FLINK-20742] Revert back to use spring-boot-maven-plugin for packaging ridesharing simulator
Before, we had to use the Maven shade plugin for packaging because we
were releasing Maven artifacts for the examples, and the
spring-boot-maven-plugin doesn't support adding custom transformers
which was required for merging NOTICE and license files for ASF legal
purposes.

Since we now no longer release these artifacts, we can revert back to
use the simple spring-boot-maven-plugin for the example.