[FLINK-21171] [python] Convert TypedValue to / from Any at Python SDK boundaries
Without reworking the whole Python SDK, we'd like to already merge the
change of replacing Any with TypedValue. To do that, we only convert
TypedValue to / from Any at the very last second before responding the
FromFunction, or at the very beginning of directly after receiving a
ToFunction.

Essentially, the SDK surface still works with Protobuf Any, but at the
protocol on-wire level, all message payloads are already TypedValues.

We will rework the whole Python SDK as a follow-up.