[FLINK-21642] Make a retry batch on recovery
According to the request-reply protocol, on recovery
we re-send the uncompleted (in-flight during a failure)
messages. But we shouldn't assume that the state
definitions are the same as in the previous attempt.
We create a retry batch and let the SDK to reply with
the correct state specs

This closes #208.