[FLINK-21678] Rename expiresAfterReadOrWrite to expiresAfterCall in Java SDK
For remote functions in StateFun, a managed remote state with Flink
expiration settings `expireAfterReadAndWrite` actually means that a
value's TTL timer will always be refreshed when attempting to invoke the
function. This is because invoking a remote function always entails
reading the state from the state backends (as defined by the request
reply protocol).

To avoid any confusion for users, we've decided to rename this
expiration configuration as `expireAfterCall` for remote functions.