[FLINK-21705] Add a SliceType
A SliceType is a type that simply treats the value as raw bytes. The
serializer implementation simply forwards the bytes as is, without any
copying. Since slices are immutable, the type is considered immutable.

This type can come in handy if users would like to avoid serialization /
deserialization in certain scenarios, e.g. sending state values as a
message to another function.

This closes #209.