[FLINK-25936][java-sdk] Adjust the ConcurrentAddressScopedStorageTest to reflect the request-reply protocol
According to the request reply protocol, in case of a missing value: a registered state
but without an assosciated value (deleted, expired, or simply never set) the invocation
will contain an associated TypedValue with the field has_value = false, and might or might not
provide the type_name field.
This commit refelects that in the test.